boot/requirements.txt
numpy
pillow
matplotlib
opencv-python
scikit-image
PyWavelets
pytest